feat: Add base code to use retrofit library
