+++ app/src/main/java/com/example/yaneodoo/RetrofitService.java
@@ -1,0 +1,9 @@
+package com.example.yaneodoo;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+
+public interface RetrofitService {
+    @GET("/users")
+    Call<String> getBistroList();
+}
